feat(transformation_page): moved the "Run all" button to header
